+++ api/src/models/auth.rs
@@ -1,14 +1,47 @@
+use jsonwebtoken::{DecodingKey, EncodingKey};
 use serde::{Deserialize, Serialize};
 
-#[derive(Serialize, Deserialize, Debug, Clone)]
-pub struct User {
-    api_key: String,
+#[derive(Deserialize)]
+pub struct AuthModel {
+    pub hash: String,
+    pub salt: String,
 }
 
-#[derive(Debug, Clone, PartialEq, PartialOrd)]
-pub enum Role {
-    DashboardAdmin,
-    BotAdmin,
+#[derive(Deserialize, Serialize)]
+pub struct LoginPayload {
+    pub api_key: String,
 }
 
-impl User {}
+#[derive(Deserialize, Serialize, Debug)]
+pub struct Claims {
+    pub exp: u32,
+}
+
+#[derive(Debug, Serialize)]
+pub struct AuthBody {
+    pub access_token: String,
+    pub token_type: String,
+}
+
+impl AuthBody {
+    pub fn new(access_token: String) -> Self {
+        Self {
+            access_token,
+            token_type: "Bearer".to_string(),
+        }
+    }
+}
+
+pub struct Keys {
+    pub encoding: EncodingKey,
+    pub decoding: DecodingKey,
+}
+
+impl Keys {
+    pub fn new(secret: &[u8]) -> Self {
+        Self {
+            encoding: EncodingKey::from_secret(secret),
+            decoding: DecodingKey::from_secret(secret),
+        }
+    }
+}
